# Optimize _dfs_convert by exiting early for common types.