--- a/changelog.txt
+++ b/changelog.txt
@@ -25,6 +25,7 @@ Changed
  - Optimized internal function ``_mkdir_p`` (#421).
  - Optimized performance of job initialization (#422).
  - Optimized performance of buffer storage (#428).
+ - Optimized performance of creating/loading synced data structures (#429).
 
 [1.5.0] -- 2020-09-20
 ---------------------