@bdice 2020-12-14
Note to self: this PR needs a few updates, along with tests/changelog. Tests are failing because of #431, which should be solved in #433.
- NumPy float64s are registered as subclasses of `float`. Need to check if that is a problem or not.
- Should the early-exit check use `isinstance(x, float)` or `type(x) == float`? Even though `isinstance` is typically preferred, it might not be the right choice here, depending on the NumPy situation previously mentioned and/or performance considerations.