@vyasr 2020-12-16
> Note to self: this PR needs a few updates, along with tests/changelog. Tests are failing because of #431, which should be solved in #433.
> 
>     * NumPy float64s are registered as subclasses of `float`. Need to check if that is a problem or not.
> 
>     * Should the early-exit check use `isinstance(x, float)` or `type(x) == float`? Even though `isinstance` is typically preferred, it might not be the right choice here, depending on the NumPy situation previously mentioned and/or performance considerations.

The issue of numpy float64s also being floats is definitely something to look into. If that's a problem, then you should use `type(x) is float` (rather than `==`). If it's not, then you'll probably want to stick with `isinstance`, you probably won't get performance benefits from directly checking types and if anything it might slow the code down.