@bdice 2020-12-16
@vyasr I agree with your comments in general. A review of the specific implementation in this PR would be helpful. The short answer on NumPy float64 was this:
- Reading data from the disk will always be type `float` (because that's how the JSON is decoded)
- Using a method like "reset state point" will use whatever type is passed. I used `type(x) in (bool, float, int, type(None), str)` because I didn't want to capture the `numpy.float64` class (I wanted to pass it through the "normal" path that will call `x.item()` to extract the Python type)